Make clippy warning-free on every target and enforce it in CI
Fix the eleven clippy warnings: checked division in the bitwise
library and the cursor extension context, collapsed ifs and a direct
iterator in the cone extractor, removed casts, a redundant closure,
expect_err, and a named row type in the tier benchmark.

The interpreter-only build exposed more: the JIT-only marshal helpers
were dead code without the feature and are now gated on it, and two
examples and one bench that call JIT APIs now declare
required-features = ["jit"] instead of failing to compile.

The all-features CI job runs clippy across all targets with warnings
as errors, with and without default features.

+++ crates/polydat/tests/examples_test.rs
@@ -233,7 +233,7 @@ fn shared_chained_hashes_differ() {
         let uid = k.pull("user_id").as_u64();
         let nidx = k.pull("name_idx").as_u64();
         let age = k.pull("account_age_days").as_u64();
-        if uid != nidx as u64 || uid != age as u64 {
+        if uid != nidx || uid != age {
             all_same = false;
             break;
         }
